feature/article/src/main/kotlin/com/example/news/feature/article/ArticleScreen.kt
package com.example.news.feature.article

import androidx.compose.foundation.ExperimentalFoundationApi
import androidx.compose.foundation.background
import androidx.compose.foundation.layout.Arrangement
import androidx.compose.foundation.layout.Box
import androidx.compose.foundation.layout.Column
import androidx.compose.foundation.layout.fillMaxSize
import androidx.compose.foundation.layout.fillMaxWidth
import androidx.compose.foundation.layout.height
import androidx.compose.foundation.layout.padding
import androidx.compose.foundation.lazy.LazyColumn
import androidx.compose.foundation.lazy.items
import androidx.compose.foundation.lazy.rememberLazyListState
import androidx.compose.material3.CircularProgressIndicator
import androidx.compose.material3.MaterialTheme
import androidx.compose.material3.Scaffold
import androidx.compose.material3.Text
import androidx.compose.runtime.Composable
import androidx.compose.runtime.getValue
import androidx.compose.ui.Alignment
import androidx.compose.ui.Modifier
import androidx.compose.ui.platform.testTag
import androidx.compose.ui.res.stringResource
import androidx.compose.ui.text.style.TextAlign
import androidx.compose.ui.unit.dp
import androidx.hilt.navigation.compose.hiltViewModel
import androidx.lifecycle.compose.collectAsStateWithLifecycle
import com.example.news.feature.article.components.NewsItem

@Composable
fun ArticleScreenRoute(
    viewModel: ArticleViewModel = hiltViewModel()
) {
    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
    ArticleScreen(
        uiState = uiState
    )
}

@OptIn(ExperimentalFoundationApi::class)
@Composable
internal fun ArticleScreen(
    uiState: ArticlesUiState,
    modifier: Modifier = Modifier,
) {
    Scaffold(
        modifier = Modifier
            .testTag(stringResource(R.string.article_screen)),
    ) { contentPadding ->
        when (uiState) {
            is ArticlesUiState.Loading -> {
                Box(
                    modifier = Modifier
                        .fillMaxSize()
                        .testTag(stringResource(R.string.article_loading)),
                    contentAlignment = Alignment.Center
                ) {
                    CircularProgressIndicator(
                        color = MaterialTheme.colorScheme.onSurface,
                    )
                }
            }

            is ArticlesUiState.Empty -> {
                Box(
                    modifier = Modifier.fillMaxSize(),
                    contentAlignment = Alignment.Center,
                ) {
                    Column(
                        verticalArrangement = Arrangement.Center,
                        horizontalAlignment = Alignment.CenterHorizontally,
                    ) {
                        Text(
                            text = stringResource(R.string.no_news),
                            style = MaterialTheme.typography.titleLarge
                        )
                    }
                }
            }

            is ArticlesUiState.Articles -> {
                val grouped = uiState.articles.groupBy { it.type }
                LazyColumn(
                    modifier = Modifier
                        .padding(contentPadding)
                        .testTag(stringResource(R.string.article_list)),
                    state = rememberLazyListState(),
                    verticalArrangement = Arrangement.Center,
                    horizontalAlignment = Alignment.CenterHorizontally
                ) {
                    grouped.forEach { (type, articles) ->
                        stickyHeader {
                            Column(
                                modifier = Modifier
                                    .height(40.dp)
                                    .fillMaxWidth()
                                    .background(MaterialTheme.colorScheme.background)
                            ) {
                                Text(
                                    text = type,
                                    modifier = Modifier.fillMaxWidth(),
                                    style = MaterialTheme.typography.headlineLarge,
                                    textAlign = TextAlign.Center,
                                    color = MaterialTheme.colorScheme.onSurface
                                )
                            }
                        }
                        items(
                            items = articles
                        ) { article ->
                            NewsItem(
                                article = article,
                                modifier = Modifier
                                    .fillMaxWidth()
                                    .padding(16.dp)
                            )
                        }
                    }
                }
            }
        }
    }
}



